pyproject.toml
[tool.poetry]
name = "django-course"
version = "0.1.0"
description = ""
authors = ["Alexandr-lab-del <gordon675843@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django-mailing = "^0.2.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
